requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# HTTP & Async
httpx==0.27.2
aiofiles==24.1.0

# AI Integrations
openai==1.51.0
google-genai==1.0.0

# Utilities
python-dotenv==1.0.1
email-validator==2.2.0
phonenumbers==8.13.47

# Caching & Rate Limiting
redis==5.1.1
slowapi==0.1.9

# Image Processing
pillow==10.4.0
python-magic==0.4.27

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2

# Development
black==24.8.0
isort==5.13.2
flake8==7.1.1
mypy==1.11.2
